Modify whd_betaiota_deltazeta_for_iota_state to not expose unfolded global fixpoints.
To not expose fixpoints, we use CClosure.whd_val instead of
Reductionops.whd_state_gen to head-betaiota normalise, relying on the
new fact that CClosure now refolds global fixpoints.

Conversely, we also add an option "expand" to force the reduction of
global fixpoints, that is to treat an applied constant hiding a
fixpoint as if already unfolded, so as to simulate the previous
behavior where whd_betaiota_deltazeta_for_iota_state was reducing
reducible Fix expressions in situations where the callers
(e.g. unification.ml) now keep the global fixpoints folded.