Modify unification.ml so that it does not expose the body of global fixpoints.
Beforehand, constants hiding fix where reduced in two steps, first
unfolding the fix then reducing the fix if ever reducible.

To avoid unfolding global fixpoint we instead add an information
"visited" which tells when a global fixpoint has already been visited
(and then unfolded in the previous way of doing).

We then trigger an head reduction of a global fixpoint when it is "visited".

Note: We also force expansion of global fixpoints in "evarconv.ml" and
it seems to work, but this is a priori a change of behavior.